# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API",
        "description": "Documentation for API.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/broker": {
            "get": {
                "tags": [
                    "Broker"
                ],
                "summary": "Get list of brokers by email",
                "operationId": "a47627622ea651354f4b33e96f998fbc",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Broker email",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "MatterNumber": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "MatterDescription": {
                                                "type": "string",
                                                "example": "Іван"
                                            },
                                            "Client": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "MatterOpened": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "CaseType": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "TitleDeedsReceived": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "MortgageOfferReceived": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "MatterCompleted": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "MatterCancelled": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "CaseHandler": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "Progress": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "AddText1": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "AddText2": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "AddText3": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "AddDate1": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            },
                                            "AddDate2": {
                                                "type": "string",
                                                "example": "ivan@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/broker/check-user": {
            "post": {
                "tags": [
                    "Broker"
                ],
                "summary": "Check if user exists",
                "operationId": "c9dd64e13067e286e0e4abbbc1cfc0d3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Evan"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "evan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}